requirements.txt
bcrypt==4.0.1
greenlet==2.0.2
mysqlclient==2.2.0
Pillow==10.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
tk==0.1.0
typing_extensions==4.7.1
